#2452 Moves redirect helper function to utils folder
