+++ e2e-tests-mocked/tests/aup/aup.spec.ts
@@ -2,63 +2,6 @@ import {expect, Page, test} from '@playwright/test';
 
 const baseURL = process.env.SBS_LOCAL_BASE_URL ?? 'http://localhost:3000';
 
-// async function mockAupApi(page: Page) {
-//   let acceptedAup = false;
-//   const jsonHeaders = { 'content-type': 'application/json', 'x-session-alive': 'true' };
-//   const currentUser = () => ({
-//     id: 999999,
-//     uid: 'urn:playwright:aup',
-//     name: 'Playwright AUP User',
-//     given_name: 'Playwright',
-//     email: 'playwright-aup@example.org',
-//     guest: false,
-//     admin: false,
-//     second_factor_confirmed: true,
-//     organisation_memberships: [],
-//     collaboration_memberships: [],
-//     service_memberships: [],
-//     user_accepted_aup: acceptedAup,
-//     CSRFToken: 'playwright-csrf-token',
-//   });
-//
-//   await page.route('**/config', route => route.fulfill({
-//     status: 200,
-//     headers: jsonHeaders,
-//     body: JSON.stringify({
-//       local: false,
-//       continue_eduteams_redirect_uri: baseURL,
-//       continue_eb_redirect_uri: baseURL,
-//     }),
-//   }));
-//   await page.route('**/api/aup/info', route => route.fulfill({
-//     status: 200,
-//     headers: jsonHeaders,
-//     body: JSON.stringify({
-//       url_aup_en: 'https://example.org/aup',
-//       url_aup_nl: 'https://example.org/aup-nl',
-//       version: 'playwright',
-//     }),
-//   }));
-//   await page.route('**/api/users/me', route => route.fulfill({
-//     status: 200,
-//     headers: jsonHeaders,
-//     body: JSON.stringify(currentUser()),
-//   }));
-//   await page.route('**/api/users/refresh', route => route.fulfill({
-//     status: 200,
-//     headers: jsonHeaders,
-//     body: JSON.stringify(currentUser()),
-//   }));
-//   await page.route('**/api/aup/agree', route => {
-//     acceptedAup = true;
-//     return route.fulfill({
-//       status: 201,
-//       headers: jsonHeaders,
-//       body: JSON.stringify({ location: `${baseURL}/welcome` }),
-//     });
-//   });
-// }
-
 test.describe.configure({mode: 'serial'});
 
 test.describe('AUP', () => {
