+++ e2e-tests-mocked/tests/organisation/organisation.spec.ts
@@ -1,47 +1,13 @@
-import {expect, Page, test} from '@playwright/test';
+import {expect, test} from '@playwright/test';
 import path from 'path';
+import {gotoWithRedirectHandler} from "../../utils/gotoWithRedirectHandler";
 
 const baseURL = process.env.SBS_LOCAL_BASE_URL ?? 'http://localhost:3000';
 const logoPath = path.resolve(__dirname, '../../../../SBS/client/src/images/surflogo.png');
 
-// Todo this could be a generic helper to handle forced redirects, like AUP
-const handleAup = async (page: Page, url: string, redirectCounter: number = 0) => {
-    const MAX_ALLOWED_REDIRECTS = 10;
-
-    if (redirectCounter >= MAX_ALLOWED_REDIRECTS) {
-        console.error(`HandleAUP - Too many redirects, trying to navigate to "${url}"`);
-        return
-    }
-
-    await page.goto(url);
-
-    await page.waitForLoadState('networkidle');
-
-    const currentUrl = page.url();
-
-    if (currentUrl !== url) {
-        if (currentUrl.includes('/aup')) {
-            const terms = page.getByText('I hereby certify that I have read the acceptable use policy and that I accept it');
-            const onwards = page.getByRole('button', {name: 'Onwards'});
-            await terms.click();
-            await onwards.click();
-
-            return handleAup(page, url, redirectCounter + 1);
-        }
-
-        if (currentUrl.includes('/missing-service-aup')) {
-            await page.locator('label[for="aup"]').click();
-            await page.locator('div.actions > button').click();
-            await page.waitForLoadState('networkidle');
-
-            return handleAup(page, url, redirectCounter + 1);
-        }
-    }
-};
-
 test.describe('Local organisation happy flow', () => {
     test('platform admin can add an organisation and see it in the list', async ({page}) => {
-        await handleAup(page, `${baseURL}/home/organisations`);
+        await gotoWithRedirectHandler(page, `${baseURL}/home/organisations`);
 
         const organisationsHeading = page.locator('.entities-search h2');
         await expect(organisationsHeading).toBeVisible();
